pyproject.toml
[project]
name = "cannabis-maturity-cv"
version = "0.1.0"
description = "Cannabis flower maturity assessment using computer vision"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "segments-ai>=1.0.0",
    "ultralytics>=8.0.0",
    "opencv-python>=4.8.0",
    "scikit-image>=0.21.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "deprecated>=1.2.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.25.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]
detectron = [
    "detectron2 @ git+https://github.com/facebookresearch/detectron2.git",
]
yolo-nas = [
    "super-gradients>=3.0.0",
]
app = [
    "streamlit>=1.28.0",
    "google-cloud-storage>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.25.0",
    "pytest>=7.4.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
